refactor(frontend): refactor sin cases faker
